refactor(zeroline-pane): move updateNameAction to wizards foundation
