+++ src/zeroline/foundation.ts
@@ -7,6 +7,7 @@ import {
   WizardActor,
   WizardInput,
   isPublic,
+  cloneElement,
 } from '../foundation.js';
 
 import { BayEditor } from './bay-editor.js';
@@ -102,26 +103,6 @@ export function getAttachedIeds(
 export type ElementEditor = Element & {
   element: Element;
 };
-
-export function updateNamingAction(element: Element): WizardActor {
-  return (inputs: WizardInput[]): EditorAction[] => {
-    const name = getValue(inputs.find(i => i.label === 'name')!)!;
-    const desc = getValue(inputs.find(i => i.label === 'desc')!);
-
-    if (
-      name === element.getAttribute('name') &&
-      desc === element.getAttribute('desc')
-    )
-      return [];
-
-    const newElement = <Element>element.cloneNode(false);
-    newElement.setAttribute('name', name);
-    if (desc === null) newElement.removeAttribute('desc');
-    else newElement.setAttribute('desc', desc);
-
-    return [{ old: { element }, new: { element: newElement } }];
-  };
-}
 
 export function cloneSubstationElement(
   editor: BayEditor | VoltageLevelEditor | SubstationEditor
