+++ test/unit/zeroline/ConductingEquipmentEditor.test.ts
@@ -1,8 +1,8 @@
 import { fixture, html, expect } from '@open-wc/testing';
 
 import { WizardInput, isCreate, isUpdate } from '../../../src/foundation.js';
-import { updateNamingAction } from '../../../src/zeroline/foundation.js';
 import { createAction } from '../../../src/wizards/conductingequipment.js';
+import { updateNamingAction } from '../../../src/wizards/foundation/actions.js';
 
 describe('ConductingEquipmentEditor', () => {
   const noOp = () => {
