+++ test/unit/zeroline/SubstationEditor.test.ts
@@ -1,8 +1,8 @@
 import { fixture, html, expect } from '@open-wc/testing';
 
 import { WizardInput, isCreate, isUpdate } from '../../../src/foundation.js';
+import { updateNamingAction } from '../../../src/wizards/foundation/actions.js';
 import { createAction } from '../../../src/wizards/substation.js';
-import { updateNamingAction } from '../../../src/zeroline/foundation.js';
 
 describe('SubstationEditor', () => {
   const noOp = () => {
